+++ history.md
@@ -1,9 +1,21 @@
+[2020/10/23]
+- Add
+  + open.commons.ssh.file.IFile: 파일/디렉토리 관련 기능 제공 
+  + open.commons.ssh.file.ListFile: 파일/디렉토리 조회 결과 객체 
+- Update
+  + open.commons.ssh.file.FileTransfer
+    - list(String): 파일/디렉토리 조회
+    - list(String, int): 파일/디렉토리 조회
+
 [2020/10/19]
 - Bugfix
   + open.commons.ssh.SshConnection
     - Session createSession(): Session 생성을 위한 Mutual Exclusive 적용
   + open.commons.ssh.file.TransferProgressMonitor
-    - 진행상태, 메시지 추가    
+    - 진행상태, 메시지 추가
+  + open.commons.ssh.exec.CommandExecutor
+    - listProcesses(int, String[]): 명령어 변경시 원본 데이터 유지
+    - listPids(int, String[]): 명령어 변경시 원본 데이터 유지
 
 [2020/10/16]
 - Snapshot:0.2.0-SNAPSHOT